Rota de listar perfil autenticado criada

+++ api/models/Cadastro.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import User
 from phone_field import PhoneField
 class CadastroModel(User):
   birth_date = models.DateField()
-  balance = models.FloatField(default=0)
+  balance = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
   cpf = models.CharField(max_length=20, null=True)
   cep = models.CharField(max_length=20, null=True)
   phone = PhoneField(null=True, blank=False, unique=True)
@@ -12,4 +12,3 @@ class CadastroModel(User):
         ('F', 'Female'),
     )
   gender = models.CharField(max_length=1, choices=GENDER_CHOICES, null=True)
-  # cpf, genero, telefone, cep
